e2e folder structure tweaks. e2e readme.
