+++ eventkit_cloud/ui/static/ui/app/tests/e2e/common/screenshot.js
