+++ eventkit_cloud/ui/static/ui/app/tests/e2e/readme.md
@@ -1,0 +1,30 @@
+# End-to-End Testing
+
+### Overview
+For end-to-end (e2e) testing, EventKit uses <a href="http://nightwatchjs.org/" target="_blank">NightwatchJS</a> as its testing framework, and <a href="https://saucelabs.com/" target="_blank">SauceLabs</a> as a cloud-based testing platform for virtualization and parallelization of tests. The e2e folder structure has been set up for easy integration of tests generated by the <a href="https://chrome.google.com/webstore/detail/snaptest/aeofjocnhdlleichkjbaibdbicpcddhp?hl=en-US" target="_blank">SnapTest</a> extension for Chrome - however, tests may still be written manually when necessary.
+
+### Setup
+In order to run e2e tests, first you need to set a couple environment variables. You'll need to set your SauceLabs credentials in `E2E_SAUCELABS_USERNAME` and `E2E_SAUCELABS_ACCESS_KEY`, and set the url you want to run the tests again in `E2E_LAUNCH_URL`.
+
+### Running Tests
+Once you have your environment variables set, you can run the current set of tests with `npm run e2e`. Any tests included in `eventkit_cloud/ui/static/ui/app/tests/e2e/tests`, `eventkit_cloud/ui/static/ui/app/tests/e2e/snaptests/tests`, or any of their subfolders will be run.
+
+### Writing Tests
+In the `eventkit_cloud/ui/static/ui/app/tests/e2e` folder, you should see three subfolders: `tests`, `common`, and `snaptests`. Hand-written tests and supporting scripts should be added to `tests` and `common` respectively. The `snaptests` directory is meant to contain only generated code from SnapTest, so none of its contents should ever be edited manually. When generating tests from SnapTest, the test code should be generated for the NightwatchJS framework in "Flat" style.
+
+### Environments
+SauceLabs supports many OS and browser options to test against. By default, tests will be run against a single environment defined in `nightwatch.json` in `test_settings.desiredCapabilities`. However, any number of extra environments may be defined in `nightwatch.json`. For example, if you wanted to define an environment using Safari 7.0 in OSX 10.9, you would need to add an object to `test_settings` like this:
+
+```json
+"<name>": {
+  "desiredCapabilities": {
+    "platform": "OS X 10.9",
+    "browserName": "safari",
+    "version": "7.0"
+  }
+}
+```
+
+The `<name>` key is completely up to you for organizational purposes. However, it would be advisable to use a naming convention - for example, `safari-7.0-osx-10.9` in this case. Possible settings for the `desiredCapabilities` fields can be found using the SauceLabs <a href="https://wiki.saucelabs.com/display/DOCS/Platform+Configurator#/" target="_blank">Platform Configurator</a>.
+
+In order to run tests against a custom environment - or against multiple environments - you can run the `e2e` script with the `-e` flag, along with the names of whichever environments you'd like to test against, separated by commas. For example, if you had two environments defined with the names `chrome-60` and `safari-7`, you could test both environments by running `npm run e2e -e chrome-60,safari-7`. This should work for any number of custom defined environments.
