+++ eventkit_cloud/ui/static/ui/app/tests/e2e/snaptests/common/components.js
@@ -1,4 +1,6 @@
-
 module.exports.bindComponents = function(browser) {
-    browser.components = {};
+
+  browser.components = {};
+
+
 }
